Fix leftover reference to the type key
If one includes the example, as written, from the docs it yields an explosion:
```
Error: The setting `type` in plugin `multiline` is obsolete and is no longer available. You can achieve this same behavior with the new conditionals, like: `if [type] == "sometype" { multiline { ... } }`. If you have any questions about this, you are invited to visit https://discuss.elastic.co/c/logstash and ask.
```